Removed some Windows-specific types
All remaining instances of UINT have been replaced with size_t.
All remaining instances of INT_PTR have been replaced with size_t (where possible) or with ptrdiff_t (identical replacement, where signed types are required) or with long long (also identical, where the StlImage helper classes use the same type) or in one case with int (in CLockList).
All remaining instances of BOOL have been replaced with bool.

+++ HITlogic/HRTImageDepthAndAngleCalculator.cpp
@@ -129,7 +129,7 @@ bool CHRTImageDepthAndAngleCalculator::givePiezoData(int /*ImageCount*/, wstring
 
 	DepthCalc.syncManualOffset(Offset);
 
-	int i = DepthCalc.getAllImageValCount();
+	int i = static_cast<int>(DepthCalc.getAllImageValCount());
 	double avgHeight = DepthCalc.getAvergaeHeight();
 	m_ImageData.setCount(i);
 	m_ImageData.setInclination(true);
